requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-cors-headers==2.4.0
django-filter==2.0.0
django==2.1
djangorestframework==3.8.2
pytz==2018.5              # via django
